Refactors button appearance theming
Consolidates button appearance theming by moving the button appearance definition into its own file.

Updates the light theme to use the new `MPButtons` struct for managing button appearances, simplifying the theme structure and improving code organization.
